+++ docs/examples/plot_quantile_regressor.py
@@ -11,8 +11,7 @@ the training objective; conformal calibration is available when
 # %%
 # Build a tiny regression panel
 # -----------------------------
-from seq_sklearn import TFTRegressor
-from seq_sklearn.config.adapters import TabularConfigParams
+from seq_sklearn import TabularConfigParams, TFTRegressor
 from seq_sklearn.data.synthetic.generator import SyntheticPanelGenerator
 
 gen = SyntheticPanelGenerator(